wowapi/requirements.txt
google-cloud-secret-manager
requests_oauthlib
